Сделаны молдели БД продуктов, Ðкатфцв

+++ src/helpers/destroy-file.ts
@@ -3,9 +3,9 @@ import uploadPath from "./upload-path";
 
 export default function (files: { path: string }[]) {
   const deleteDir = (filePath: string[]) => {
+    if(!filePath) return;
     const dir = JSON.parse(JSON.stringify(filePath))
     dir.pop()
-    console.log(dir)
     if (dir.length === 0) return
     fs.readdir(uploadPath(dir), (err, files) => {
       if (files.length === 0) {
@@ -16,11 +16,13 @@ export default function (files: { path: string }[]) {
     });
   }
   files.map((file) => {
+    if(!file.path) return;
     const filePath = uploadPath([file.path])
     
     fs.unlinkSync(filePath);
   })
   files.map((file) => {
+    if(!file.path) return;
     const filePath = file.path.split('\\')
     
     deleteDir(filePath)
